+++ src/General/Configs.go
@@ -24,17 +24,11 @@ type Configs struct {
 
 func GetConfig() (*Configs, error) {
 	data, err := os.ReadFile("config.yaml")
-	if err != nil {
-		LogErr(err.Error())
-		return nil, err
-	}
+	if err != nil { LogErr("GetConfig: " + err.Error()); return nil, err }
 
 	var cfg Configs
 	err = yaml.Unmarshal(data, &cfg)
-	if err != nil {
-		LogErr(err.Error())
-		return nil, err
-	}
+	if err != nil { LogErr("GetConfig: " + err.Error()); return nil, err }
 
 	return &cfg, nil
 }