+++ src/Registration/Reg.go
@@ -26,32 +26,21 @@ func RegisterNewAgent(agentSymbol string) string {
 	}
 
 	jsonBytes, err := json.Marshal(payload)
-	if err != nil {
-		General.LogErr(err.Error())
-	}
+	if err != nil { General.LogErr("RegisterNewAgent: " + err.Error()) }
 
 	req, _ := http.NewRequest(http.MethodPost, "https://api.spacetraders.io/v2/register", bytes.NewBuffer(jsonBytes))
-	if err != nil {
-		General.LogErr(err.Error())
-	}
+	if err != nil { General.LogErr("RegisterNewAgent: " + err.Error()) }
 
 	req.Header.Add("Authorization", "Bearer "+CFG.Tokens.Account)
 	req.Header.Add("Content-Type", "application/json")
 
 	client := &http.Client{Timeout: 10 * time.Second}
 	resp, err := client.Do(req)
-
-	if err != nil {
-		fmt.Fprintf(&returns, "Error performing request: %v\n", err)
-		return returns.String()
-	}
+	if err != nil { General.LogErr("RegisterNewAgent: Error performing request:" + err.Error()); return returns.String() }
 	defer resp.Body.Close()
 
 	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		fmt.Fprintf(&returns, "Data Read Failed: %s", err)
-		return returns.String()
-	}
+	if err != nil { General.LogErr("RegisterNewAgent: Data read failed:" + err.Error()); return returns.String() }
 
 	fmt.Fprintf(&returns, "%s", body)
 	return returns.String()